requirements.txt
anyio==4.4.0
asgiref==3.8.1
channels==4.1.0
click==8.1.7
colorama==0.4.6
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
h11==0.14.0
httptools==0.6.1
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
whitenoise==6.7.0
